Add files via upload
The attached file has a python code which is a simple calculator which can perform basic arithmetic operations. It accepts two user inputs and an operator and performs the calculation and gives the result. As it is a simple calculator it can perform only 4 arithmetic operations like addition, subtraction, multiplication and division